Change range type to float

+++ src/Shared/Domain/Exception/InvalidRangeException.php
@@ -7,9 +7,9 @@ namespace Teamleader\Discounts\Shared\Domain\Exception;
 abstract class InvalidRangeException extends DomainException
 {
     public function __construct(
-        protected readonly int $value,
-        protected readonly int $min,
-        protected readonly ?int $max = null
+        protected readonly float $value,
+        protected readonly float $min,
+        protected readonly ?float $max = null
     ) {
         parent::__construct();
     }
